Fix tone() output for pinnumbers beyond 8

+++ cores/arduino/Tone.cpp
@@ -18,7 +18,7 @@ static HardwareTimer TimerTone(TIMER_TONE);
 static void tonePeriodElapsedCallback()
 {
     uint32_t port = gpio_port[GD_PORT_GET(TimerTone_pinInfo.pin)];
-    uint8_t pin =  gpio_pin[GD_PIN_GET(TimerTone_pinInfo.pin)];
+    uint32_t pin =  gpio_pin[GD_PIN_GET(TimerTone_pinInfo.pin)];
 
     if (port != (uint32_t)NC) {
         if (TimerTone_pinInfo.count == -1) {
